Refactor utilities
Moved pkg/syncer/broker/utils -> pkg/util for general use.

Moved helper functions in broker_suite_test to pkg/syncer/test for
general use.
