+++ pkg/syncer/broker/config.go
@@ -16,7 +16,7 @@ type brokerSpecification struct {
 	Ca              string
 }
 
-func buildConfig() (*rest.Config, string, error) {
+func buildBrokerConfig() (*rest.Config, string, error) {
 	brokerSpec := brokerSpecification{}
 	err := envconfig.Process("broker_k8s", &brokerSpec)
 	if err != nil {
